// src/app/login/Models/UserRegistration.ts
export interface UserDetails {
    FirstName: string;
    LastName: string;
    Email: string;
    MobileNo: string;
    Gender: boolean;
    Address: string;
    UserName: string;
    Password: string;
    ConfirmPassword: string;
    Seq1: string;
    SeqAns1: string;
    Seq2: string;
    SeqAns2: string;
    Seq3: string;
    SeqAns3: string;
}